Create README.md
ASSIGNMENT V

DEADLINE 15th February,2017

Function

Write a Program in C to input a number and find whether the number is exactly divisible by 5 and 7 or not using functions:

No return no argument

No return with argument

Return with no argument

Return with argument

WAP in C to input a number and check whether the number is palindrome or not using function.

WAP in C to find whether the given number is prime or not using function
WAP in C to print all prime number less than 500 using function
WAP in C to print print TWIN PRIME less than 500 using function. [if two consecutive odd numbers are both prime then they are known as twin prime].
WAP in C to convert decimal number to binary using functions
WAP in C to find all the divisor of a given number using functions
Write a function Power that computes X raised to the power Y for integer X and Y and return double type value.
Arrays and Strings

Write a Program in C to input any 10 elements in an array and display them using function. (Passing individual array element).

Write a program in c to input any 10 number in an array and sort them in ascending order using function. (Pass Whole array).

Write a program in c to input any 10 number in an array and sort them in descending order using function. (Pass Whole array).

Recursion

Write a program in C to find factorial of a given number using Recursive Function
WAP in C to generate the Fibonacci series using recursive Function.
Write a recursive function to generate the particular term in Fibonacci Series. ( eg. 8th term: 13).

WAP in C to display the reverse number using Recursion.

WAP to find the sum of digits entered by the user.

WAP to find the sum of Natural number up to 10 using recursion.

WAP in C to ask a number and Power to be calculated for that number.